refactor: вынес подключение к Redis в отдельный файл
